# docs/swagger.yaml
definitions:
  dto.Pageable:
    properties:
      offset:
        type: integer
      page:
        type: integer
      size:
        type: integer
    type: object
  dto.Sortable:
    properties:
      direction:
        type: string
      field:
        type: string
    type: object
...
  request.AccountRequest:
    properties:
      id:
        type: integer
      jenis:
        type: string
      karyawan:
        $ref: '#/definitions/request.IdRequest'
      nama:
        type: string
      rekening:
        type: string
    required:
    - jenis
    - nama
    - rekening
    type: object
  request.DetailKaryawan:
    properties:
      id:
        type: integer
      nik:
        type: string
      npwp:
        type: string
    required:
    - nik
    - npwp
    type: object
  request.EmployeeRequest:
    properties:
      alamat:
        type: string
      detailKaryawan:
        $ref: '#/definitions/request.DetailKaryawan'
      dob:
        type: string
      id:
        type: integer
      nama:
        type: string
      status:
        type: string
    required:
    - alamat
    - dob
    - nama
    - status
    type: object
...
  request.IdRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  response.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.PaginationData:
    properties:
      content: {}
      emptyPage:
        type: boolean
      firstPage:
        type: boolean
      lastPage:
        type: boolean
      numberOfElements:
        type: integer
      pageable:
        $ref: '#/definitions/dto.Pageable'
      sortable:
        $ref: '#/definitions/dto.Sortable'
      totalElements:
        type: integer
      totalPages:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/idstar/karyawan/{id}:
    get:
      consumes:
      - application/json
      operationId: GetEmployeeById
      parameters:
      - description: Karyawan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Karyawan By Id
      tags:
      - karyawan
  /v1/idstar/karyawan/delete:
    delete:
      consumes:
      - application/json
      operationId: DeleteEmployee
      parameters:
      - description: Delete Karyawan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.IdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Karyawan
      tags:
      - karyawan
  /v1/idstar/karyawan/list:
    get:
      consumes:
      - application/json
      operationId: GetEmployeeList
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Field
        in: query
        name: field
        type: string
      - description: Direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginationData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Karyawan List
      tags:
      - karyawan
  /v1/idstar/karyawan/save:
    post:
      consumes:
      - application/json
      operationId: CreateEmployee
      parameters:
      - description: Create Karyawan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Karyawan
      tags:
      - karyawan
  /v1/idstar/karyawan/update:
    put:
      consumes:
      - application/json
      operationId: UpdateEmployee
      parameters:
      - description: Update Karyawan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Karyawan
      tags:
      - karyawan
  /v1/idstar/rekening/{id}:
    get:
      consumes:
      - application/json
      operationId: GetAccountById
      parameters:
      - description: Rekening ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Rekening By Id
      tags:
      - rekening
  /v1/idstar/rekening/delete:
    delete:
      consumes:
      - application/json
      operationId: DeleteAccount
      parameters:
      - description: Delete Rekening Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.IdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Rekening
      tags:
      - rekening
  /v1/idstar/rekening/list:
    get:
      consumes:
      - application/json
      operationId: GetAccountList
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Field
        in: query
        name: field
        type: string
      - description: Direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginationData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Rekening List
      tags:
      - rekening
  /v1/idstar/rekening/save:
    post:
      consumes:
      - application/json
      operationId: CreateAccount
      parameters:
      - description: Create Rekening Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Rekening
      tags:
      - rekening
  /v1/idstar/rekening/update:
    put:
      consumes:
      - application/json
      operationId: UpdateAccount
      parameters:
      - description: Update Rekening Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Rekening
      tags:
      - rekening
...
swagger: "2.0"